+++ apps/user-manager/src/main/java/com/warlock/user/model/UserToken.java
@@ -1,0 +1,13 @@
+package com.warlock.user.model;
+
+import lombok.Builder;
+import lombok.Data;
+
+import javax.validation.constraints.NotNull;
+
+@Data
+@Builder
+public class UserToken {
+    @NotNull
+    private String token;
+}