+++ apps/user-manager/src/test/java/com/warlock/user/controller/UserControllerTest.java
@@ -1,37 +1,28 @@
 package com.warlock.user.controller;
 
-
-import com.warlock.user.model.User;
-import org.junit.jupiter.api.Test;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.warlock.user.service.UserService;
+import io.jsonwebtoken.JwtBuilder;
+import io.jsonwebtoken.JwtParser;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.web.servlet.MockMvc;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-@SpringBootTest
+@WebMvcTest(UserController.class)
 public class UserControllerTest {
     @Autowired
-    UserController userController;
+    MockMvc mockMvc;
 
-    @Test
-    public void login_ReturnsSuccess() {
-        var response = userController.login();
-        assertTrue(response.hasBody());
-        assertThat(response.getBody()).isNotNull();
-        var responseBody = response.getBody();
-        assertTrue(responseBody.containsKey("token"));
-        assertTrue(responseBody.containsKey("user"));
-        var token = (String) responseBody.get("token");
-        assertThat(token).contains("CHUMP");
-        var user = (User) responseBody.get("user");
-        assertThat(user.getUsername()).isEqualTo("hrothgar.warlock");
-        assertThat(user.getId()).isEqualTo("1a");
-        assertThat(user.getPassword()).isEqualTo("p4ssw0rd");
-        var profile = user.getProfile();
-        assertThat(profile.getId()).isEqualTo("2b");
-        assertThat(profile.getBirthdate()).isEqualTo("1981-06-18");
-        assertThat(profile.getFirstName()).isEqualTo("Hrothgar");
-        assertThat(profile.getLastName()).isEqualTo("Warlock");
-    }
+    @Autowired
+    ObjectMapper mapper;
+
+    @Autowired
+    JwtParser jwtParser;
+
+    @Autowired
+    JwtBuilder jwtBuilder;
+
+    @MockBean
+    UserService service;
 }